src/scene_worker/requirements.txt
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.2
azure-identity==1.16.1
azure-storage-blob==12.22.0
azure-servicebus==7.12.2
httpx==0.27.2
tenacity==8.2.3
Pillow==10.2.0
supabase==2.9.0
